+++ src/BaseModel.php
@@ -4,17 +4,15 @@ declare(strict_types=1);
 
 namespace Inspirum\Arrayable;
 
-use JsonSerializable;
-use Stringable;
 use function json_encode;
 use const JSON_THROW_ON_ERROR;
 
 /**
  * @template TKey of array-key
  * @template TValue
- * @implements \Inspirum\Arrayable\Arrayable<TKey, TValue>
+ * @implements \Inspirum\Arrayable\Model<TKey, TValue>
  */
-abstract class BaseModel implements Arrayable, JsonSerializable, Stringable
+abstract class BaseModel implements Model
 {
     /**
      * @return array<TKey, TValue>