+++ CHANGELOG.md
@@ -6,7 +6,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 
-## [Unreleased](https://github.com/inspirum/arrayable-php/compare/v1.0.0...master)
+## [Unreleased](https://github.com/inspirum/arrayable-php/compare/v1.1.0...master)
+
+
+## [v1.1.0 (2022-07-22)](https://github.com/inspirum/balikobot-php/compare/v1.0.0...v1.1.0)
+### Added
+- Added [**Collection**](./src/Collection.php) interface
+- Added [**Model**](./src/Model.php) interface
 
 
 ## v1.0.0 (2022-07-04)